# perf: slightly improve parsing a port