Before:

```
test port           ... bench:         222 ns/iter (+/- 5) = 108 MB/s
```

After:

```
test port           ... bench:         211 ns/iter (+/- 7) = 113 MB/s
```